requirements.txt
compas_assembly >=0.7.0
compas_cra
compas_dr
compas_fd
compas_gmsh
compas_tna
compas_ifc
compas_model
compas_viewer
compas_notebook
compas_session >=0.4.5
compas_rv
git+https://github.com/BRG-research/compas_grid.git
git+https://github.com/BlockResearchGroup/compas-Masonry.git